Update patch set 1
Patch Set 1:

"but why is bashate not running on these files..."

I wondered the same too so I tried it by hand and if the version on pypi is being used, no errors are produced on either file. Using the version available from github there are errors produces but not for the problem fixed in this commit.

So something with expectations and reality is not quite right...

Patch-set: 1