Update patch set 1
Patch Set 1: Code-Review-1

It's kinda unclear to me why this is needed.  Presumably you're still using the deploy kernel and ramdisk that devstack is building and publishing, why wouldn't you also just be using the cirros image that devstack publishes as well (this is the "user" image)  Populating glance with your own custom images seems more like a post-deployment task that a user or administrator would take care of after running devstack, not something devstack should be responsible for.

That said, I could imagine devstack providing a way of specifying one or more custom user images that it would publish into glance in addition to the stuff it already uploads.. But I think that should be something that could be consumed more broadly and not only in the ironic case.

Patch-set: 1
Reviewer: Gerrit User 1420 <1420@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1