Update patch set 2
Patch Set 2:

More: For sake of completeness I tried yesterday (before I knew you were doing this) to do these changes by hand, but doing it in one fell swoop resulted in the same problem as the one in my previous comment.

Most of the changes are just fine, but there are a few places where nesting and quoting appear to break stuff in unexpected ways.

Patch-set: 2