Update patch set 1
Patch Set 1:

(2 comments)

@Jordan thanks for review, I wasn't too happy with the code duplication myself but got distracted by the need to call out different keystone parameters (btw, there is already some precedent for that in lib/heat [1]). I'm just testing a new version which handles the domain option in the existing functions.

The current running tests verify swift access control for users (1 & 3) in the same project (1), and for users (1 & 2) in different projects (1 & 2), all in the default domain. 

The currently skipped tests verify access control for users (1 & 4) in different projects (1 & 4) in different domains, in which case swift imposes different conditions on access, hence the need for a fourth user in the new domain.

[1] https://github.com/openstack-dev/devstack/blob/master/lib/heat#L267

Patch-set: 1