Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Do we really need pbr?  I mean, the overhead for a not-a-Python-project is getting...well...very Python-y

tools/build_docs.sh does the work of generating the script pages and can't go away yet.  I'll clean it up to remove the old publishing bits, but it needs the sphinx build added to it.

Patch-set: 2
Label: Code-Review=-1