Require distro-provided python-requests for SSL/TLS proxy
requests provides a method, where(), which returns the location
of the CA bundle. One is provided with the package but it
recommends that distros override this. Both Fedora and Ubuntu
do this, and provide a means of installing additional CAs into
the bundle. This allows devstack to add its own self-signed CA
to the global bundle. Not having this causes subtle failures.

Change-Id: Ib976c3a53344717b9e9551acd16d3402c48cbd64