Remove duplicate HEAT_CFN_IMAGE_URL assignment
There are two places were the URL is assigned within stackrc.
It is impossible to override one of the assignments from localrc.
We want to be able to override it because, in our case, we can not
use https to retrieve the image.

With this commit we allow the URL to be configured and remove
redundant assignment.

Change-Id: I8e5ae77644cec2b87185d740505bc2bbce9c89ce