Add support for tuskar-api service
Adds initial support for configuring and starting the tuskar-api
service, when "t-api" is added to the localrc.

To enable add "enable_service t-api" to your localrc, this will
install tuskar, python-tuskarclient, and start the service.

The aim of this addition is to provide a more accessible (non devtest)
way for developers to run up tuskar in a familiar devstack environment.

See the official repos for more information:
https://github.com/openstack/tuskar/
https://github.com/openstack/python-tuskarclient

Change-Id: Id0c3c0d3a38100c66dbe6e3adf1f715162f99742