Fix expansion of variables in localrc
Use a trick of passing the incoming value through an eval'd "cat" run
to expand any variables within it, whilst retaining the quoting.  Idea
cribbed from an example in [1].

The movie "Inception" had less levels of nesting than this ini setting
stuff ... I'm about as sure of the reality of variables passing
between each level as the guy with his spinning top.  However,
test-cases are added.

[1] http://stackoverflow.com/a/20316582

Change-Id: I9fc99f3716cc53366907878adb00ae6cf3898f14
Closes-Bug:#1386413