Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

My main concern here is things like neutron that can use sudo *a lot* in the course of operation.  this seems like it could cause this to both overflow and become fairly useless

maybe if it was limited to the stack user's command only?

Patch-set: 1
Label: Code-Review=-1