Update patch set 1
Patch Set 1: Code-Review-1

I don't think this is the way we want to solve this because we wanted to get rid of the distro version binding on requests so we could test newer versions in devstack. 

Instead we should make it so that the CA is registered into requests lib.

Patch-set: 1
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1