Update patch set 1
Patch Set 1:

I think the resulting code might be even more unreadable.

We can't easily twist the existing tools to use a different bundle location. In both Ubuntu and Fedora the general manage a bundle tools have hardcoded paths all over.

The bundle is just a huge text file consisting of concatenated PEM files. Great care would be needed to a) not corrupt and b) ensure uniqueness when adding/removing the devstack CA certificates. It might be possible to add some sort of marker to avoid having to actually parse the PEM files but yikes.

Patch-set: 1