Create validate step for Neutron third-party programs
Add a new function to lib/neutron that can be implemented by scripts under
lib/neutron_thirdparty to validate that the current configuration
is compatible with the third-party program. This can be used, for example, to
ensure that enabled services do not conflict with the third-party program
and that pre-requisite services are enabled. Adding a validate step
relatively early in the stacking process allows it to fail fast when
incompatible configuration is detected.

Call this function in a new section of stack.sh, 'Validate Configuration',
directly after the 'Configure Projects' section.

Also, fix a typo in lib/neutron_thirdparty/bigswitch_floodlight

Change-Id: Icf58b6bec0f1f31320b4e126d6e27a047e31da74
Closes-Bug: #1386267