Update patch set 7
Patch Set 7:

> so I think for something like tuskar I'd rather not put the function contents in lib/tuskar, but just in 70-tuskar.sh itself. Mostly leave the lib/ for integrated release pieces, but let extras.d be available to anyone.

Can you clarify what you hope to achieve by my doing that, is that the plan for everything long term, or is tuskar being given special treatment for some reason?

My main objection is that it's less familiar for developers working on devstack, as *everything* else, including incubated projects has their function library under lib, in a standard format.

If I go ahead and mangle all that into extras.d, tuskar will be the only one with all the library code in a different place, so I'd rather not do that unless it's planned to be a new convention tbh.

That said, if all projects did this, it would fix the slightly weird chicken/egg problem which exists with the two enable_service lines which are needed so both the source step and subsequent install recognise the service is enabled.

Patch-set: 7