Update patch set 1
Patch Set 1:

This appears to have worked, you can see in the pip-freeze that the python-fooclients are installed from git, as expected.

The fact that these tests are this way, and the patches below them have pip versions of fooclients, means that the code is working as expected now.

Patch-set: 1