Update patch set 1
Patch Set 1: Code-Review-1

Whilst I agree this is suitable for Master, I disagree this is appropriate for backport.  Although we shouldn't be installing this from apt, we have.. This means that it has caused the situation of an undocumented requirement, which is version bound by deployers based on how they deploy (ie, it works for them).

We do not know which versions work, so feel we should leave this as-is for the current stable/* branches.

Patch-set: 1