Update patch set 1
Patch Set 1:

I'm unconvinced that's a reason to block this change. The package was being preinstalled so that it could be patched to provide a workaround for, e.g., tempest's use of testtools. I found no direct uses of unittest2/discover anywhere in devstack itself, and the stated workaround is also no longer necessary. You could counterargue that until we get devpy implemented, we should install packaged versions of all tempest's dependencies instead to help improve performance when it's run.

Patch-set: 1