Update patch set 1
Abandoned

At this point it appears that providing a correct minimum requirement for unittest2 in testtools is causing tempest to successfully pull in a newer version from PyPI rather than relying on the preinstalled system package, so no need to rush in this one change. Instead it can be rolled up into a larger cleanup effort of packages within devstack if someone wants to mount that (looks like Monty is already taking a stab at it).

Patch-set: 1
Status: abandoned