Update patch set 1
Patch Set 1: Code-Review-1

I agree that things tend to work inspite of this, rather than because of it

However, if this is the new policy we're following, I'm a little worried that we'll end up making *really* hard to debug when there's conflicts between what other parts of the system require and what we've installed.

For example, I7ef0fa35c650e323a8945ed7c5c7754ca62a0ca5 was an absolute nightmare to figure out.  

Maybe I'm arguing that we do this, but require working in a virtualenv.  Is there history behind why we don't do this?

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1