Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Typo in the commit message but otherwise looks good.

I wonder about the impact of this. I think it is the right way for the world to be, but making this change will mean the distro packages (no longer being used here) are exercised against openstack-at-large much less. So by making the change we're making a statement about what devstack is and is for. Is that on purpose?

Patch-set: 1
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1