Update patch set 4
Patch Set 4:

Latest patchset addresses comments from Kyle and Dean.

Chris Dent, this guide currently does not address "Joe Developer" who wishes to do the network configuration that you are requesting. At this point, this is just the documentation that I have tried to put together from a physical lab that I run devstack on, as well as a Vagrant project (https://github.com/bcwaldon/vagrant_devstack) that can be used to build a DevStack instance with Neutron networking enabled.

I tried to re-work the Vagrant project to use a single interface - but I ran up against a couple routing issues and other miscellanea that turned the whole thing into a large rabbit hole. At this point, I'd like to try and contribute this doc and maybe work towards something that addresses your concerns in the future.

Patch-set: 4
Label: Verified=0