Add data-processing service for Sahara
When registering endpoint with data_processing Keystone transforms it
into data-processing. This problem causes sahara to not find the
endpoint afterwards

We need to have two endpoint for correct working tempest and horizon
with devstack. After resolving bug in tempest and horizon need to remove
old data_processing endpoint

Change-Id: I97827d23ffe8a1218abd61e76804b918b1b7cbe0
Partial-bug: #1356053