Update patch set 5
Patch Set 5:

I'm happy for the instructions for joe developer to be left for another time, but since this is an entirely new files with the name "neutron" it looks like it should be more than it actually is. Would it make sense to leave some "to be done" sections in the document so that people know the specifics contexts in which this information does and does not apply?

Patch-set: 5
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0