Update patch set 2
Patch Set 2: Code-Review-1

yeah, don't want to reown /etc :)

i think this is on the right track.  why no keep this in TEMPEST_CONFIG_DIR? i suggest create it in there, and symlink in to /etc

IMO it's ok to assume devstack owns this config file.  if people need extra things in there, it should be worked into the devstack framework for consistency

Patch-set: 2
Label: Code-Review=-1