Add compute use_json_api_for_all_tests option
Now Nova API XML support is deprecated and we will drop it in Kilo.
And we will use Nova JSON clients for the other projects' tests on
Tempest as the default for master branch.

On the other hand, we need to keep the test coverage for stable
branches. This patch adds a option to keep using Nova XML clients
for the other projects' XML tests.

This option use_json_api_for_all_tests is added by the Tempest commit
Ic0540beb8ec0c5c1ce549ba46dc52b1870407665 . So we need to wait for the
Tempest commit before this.

Change-Id: I5aa500dd2717b7edd628fa45590c62e754a432d0