Update patch set 3
Patch Set 3: Code-Review-1

this seems incredibly convoluted.

If we want the ability to set this via tempest we should create a boto.cfg that tempest carries around, set BOTO_CONFIG in the tox env for tempest to use that, then just do the iniset for the ca bundle.

Using existing boto.cfg on the system is just going to lead to bugs.

Patch-set: 3
Label: Code-Review=-1