Install pkg python-memcache for Nova
As the package "python-memcache" is not installed by DevStack,
it causes "ImportError" when this package is imported in the module
"nova.openstack.common.memorycache".

This patch is to force DevStack to install the pkg
"python-memcache" for Nova by default. This installation should not
be put under any condition as the condition of using this
pkg is controlled by Nova source code.

Change-Id: I486cb6cb23a97ed9dd9f24d75c4600395fae8be8
Closes-Bug: #1394448