Allow for pre-created Ironic libvirt VMs
If using the pxe_ssh driver, the only VMs that may be used for testing are
those auto-created locally by devstack.  This adds an additional option to
pass a file to devstack that contains required information for local or
non-local libvirt virtual machines to be enrolled and used with the pxe_ssh
driver.

Change-Id: I01e13e01a0bcd2cc2d6658d6b5966cd4956e2d8c