Update patch set 2
Patch Set 2: Code-Review-1

If this needs to be altered for the tests, please change this in the test workflow for that particular datastore. Please do not update the devstack defaults for all datastores in this case.

Patch-set: 2
Reviewer: Gerrit User 5293 <5293@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1