Add TEMPEST_COMPUTE_CATALOG_TYPE
Now Nova contains both v2 and v2.1 APIs as its REST API, and we are
testing v2 API only on the gate. To test v2.1 API also, this patch
add a new item TEMPEST_COMPUTE_CATALOG_TYPE which allows us to switch
a compute service catalog type. The default value is for testing v2
API and we will be able to test v2.1 API by setting "computev21" as
the value.

Change-Id: Iaaa530bda6bcdae75e86be8dbe572abe4396e8e9