Update patch set 1
Patch Set 1: Workflow-1

Going to WIP for now.  This seems to have exposed the need for floating ip pools to be defined at both the child and parent level for all functionality to work.  Different things work based on where they're defined, and it's looking like we may want to keep it in the child cell for now.

Ideally the issues here would get fixed so this could be merged.  But since we're deprecating this cells functionality in favor of a new solution I'll likely look at skipping tests for functionality that needs a lot of work to fix.

Patch-set: 1
Reviewer: Gerrit User 5441 <5441@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1