Update patch set 10
Patch Set 10: Code-Review+1

OK, whatever on the location.  Go with what Sean wants.

I don't have any more issues with the code for what it is.

Overall, I'm a little uneasy this.  We've come this far, but I would be very negative towards further expansion of other projects in this direction because I feel it rolls too much maintenance burden into the core of devstack.  This is really a whole set of separate scripts that has been wedged in behind a config variable -- it would be just as valid to separate all this setup stuff out into some sort of contrib type area and just have devstack know to leave the system alone and use the qemu/libvirt already installed for it.

Patch-set: 10
Label: Code-Review=+1