Hard code icehouse service extension lists
This commit adds the hard coded service extension lists for icehouse
in the tempest config files. To enable gating with tempest master on
all supported branches the extension list has to be explicit on stable
branches to ensure we don't attempt to run tests for new extensions on
stable branch services. This commit adds these lists for the icehouse
branch.

Implements blueprint branchless-tempest-extensions

Change-Id: I14b4301f9211529e5ff6fe56f65392c1d198a18a