Hard code juno service extension lists
This commit adds the hard coded service extension lists for juno
in the tempest config files. To enable gating with tempest master on
all supported branches the extension list has to be explicit on stable
branches to ensure we don't attempt to run tests for new extensions on
stable branch services. This commit adds these lists for the juno
branch.

Implements blueprint branchless-tempest-extensions

Change-Id: Id252581f180045e265d7163bc236ce2c76d40da8