Update patch set 3
Patch Set 3: Code-Review-1

adding the service seems fine

but all those config options; they aren't transformational and can just be passed through from local.conf and we don't need to carry code for them

i know it's the way it's always been done, but in the past the documentation was hard to edit and non-standard.  now it's all rst just like everything else.

I think this presents a fine opportunity to create a doc/source/guide/nova.rst.  it's first entry can be a note about the n-serial service, what it does, and that options can be set via local.conf entries with maybe an example.

Patch-set: 3
Label: Code-Review=-1