Update patch set 1
Patch Set 1:

The grenade jobs failed because of a "bug" in the grenade: the  target tempest uses images of base tempest. And since the patchset upgrades Cirros version the target tempest expects to have Cirros 0.3.3 but it has Cirros 0.3.2 . As a result some tests are failing.
In my opinion it is a trivial bug because it has a workaround and it can be reproduced only here.

Patch-set: 1