Install Juno RDO repos for RHEL7
RHEL7 does not include Open vSwitch in the default repos, but it
is available via the RDO repo.  This patch automatically
configures the RDO repo for RHEL7.

Move this into the existing rhel6/rhel7 section.  We update RHEl6 to
the latest icehouse release, but it is not supported with Juno

Change-Id: I8f5e49f8b8012a6ceaa3e087db213ce902467111