Update patch set 5
Patch Set 5:

Hi Zhang and Yamamoto,

Please look again at the following part extracted from the script lib/keystone of DevStack:

  elif [[ "$KEYSTONE_TOKEN_BACKEND" = "memcache" ]]; then
    iniset $KEYSTONE_CONF token driver   
    keystone.token.persistence.backends.memcache.Token

It should be clear to see that the pkg "memcache" here is not the python-memcache pkg, but it is a Keystone-specific pkg whose implementation can be anything.

We do really need to have DevStack install the standard pkg named "python-memcache". For detail, please read the "Bug description" on this link: https://bugs.launchpad.net/devstack/+bug/1394448

Hope to convince both of you!

Patch-set: 5