Update patch set 5
Patch Set 5:

Yamamoto,

The pkg "memcache" of Keystone is not the standard pkg "python-memcache" needed by Nova source code as specified.

I used to get the error of "Import Error" due to this missing pkg. That's why I reported this bug and hope the patch will be passed through so that no one else will encounter same trouble as I did

Patch-set: 5