Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

This did turn out a lot cleaner than I expected.

The way the settings source works, it might be nice to have that for all repos.  But I see where that could be abused too, but opening it up to external repos and not the main repos seems backwards maybe?

Patch-set: 2
Label: Code-Review=-1