Configure auth_token middleawre by auth plugin
As of release 1.3 auth_token middleware can be configured to use any
authentication plugin. This allows us to move to the more generic
password mechanism which will default to using keystone v3 if available.

This will allow in future revisions to move the devstack service users
out of the default domain.

Change-Id: Ieac26806bd420aa08fc79bbc6a11eb6a1c15c7df