Update patch set 7
Patch Set 7:

Yamamoto,

Jenkins failures are all invalid and irrelevant to this patch. Please look at the failure content of Jenkins to know why!

Looking at the code snippet given in the "Bug description", we can see that the code line "import memcache" is under one condition that's controlled by the caller of the function. It's controlled inside source code. That's why we have to provision the installation of this pkg.

Do you agree?

Patch-set: 7