Update patch set 3
Patch Set 3:

Can we consider that for a follow-on patch? This patch attempts to get the final neutron split stuff working, and changing that token will require a lot of churn and some care that we preserve backwards compatibility. I mean, we still use 'q' for neutron, and the vpn token is an issue?

Patch-set: 3