Allow the option of not starting openstack services
The functional test job of some openstack projects (e.g. neutron) uses devstack
to set up the environment for the functional tests. This currently involves
running devstack to get the service installed, and then running unstack because
the tests do not want the service actually running.

Thus it would be convenient if we had a way to tell devstack to stop after
installing the enabled services, and not actually start them.

We add a configuration option START_OPENSTACK_SERVICES which defaults to True
to preserve existing behavior. Setting this option to False will prevent
OpenStack services from starting.

Re-organize some config_, init_, and start_ functions to simplify the flow
where the option is checked. Also add a post-start phase for services, called
"commission_" where setting up of resources is done by some services.

Change-Id: I867ef1362543a28a9031df1df35f254b02e377bb
Closes-bug: #1405838