Update patch set 1
Patch Set 1:

I figured out what was wrong with the read permissions and why. I submitted patch to tempest as https://review.openstack.org/#/c/144534/ if anyone can have a look on it and review I will be thankful.
Also not sure how to put dependency between this review and tempest change 144534.

Patch-set: 1