Update patch set 2
Patch Set 2:

What I tried to do was to use the same technique to skip the post-startup commands as for skipping the startup commands (compare function commission_openstack_service to function start_openstack_service). I not sure if it is 'necessary', but I felt it turned out looking a little neater.

I am also quite uncertain about the choice of the verb 'commission'. I could not find a better alternative and am open to other suggestions.

Patch-set: 2