Update patch set 2
Patch Set 2: Code-Review-1

I have a crazy idea - let's not do this by modifying devstack (at least not right now).  You've already done it for the db setup of the functional job - sourced devstack and ran the necessary functions to install the db.  Well, in the gate_hook we have sudo privileges until we run devstack-vm-gate.sh and it removes them, so we should be able to first invoke the devstack function to install the rootwrap function (and whatever else has to happen before sudo is revoked).  Voila, a devstack vm with only mysql and rabbit running!

Patch-set: 2
Label: Code-Review=-1