Ensure also last extension in the list can be removed
The algorithm for removing disabled extensions from the list which
will be supplied to tempest assumes that an extension name is
followed by a comma. This is not true for the last element in the
list; the current algorithm cannot therefore remove it.

This patch fixes this issue by adding a comma at the end of the
extension list string. The additional comma is removed immediately
before returning the string.

Change-Id: I01cec39dae8c6a99620d43471c9a7d34b84da846