Update patch set 7
Patch Set 7:

Also it appears additional configuration is necessary for any record meters to made available to ceilometer. Is that coming later or is it expected?

When you create a devstack without gnocchi, but with ceilo support, immediately after the stack.sh complete `ceiometer meter-list` will contain some storage and image meters. When configuring with gnocchi that doesn't happen. Presumably this is because dispatch to gnocchi is not happening. How do we turn that on?

Patch-set: 7