Update patch set 3
Patch Set 3:

Hi Hirofumi,
The default config option is assigned when the config variable is set on the l3_agent.ini file.
But if the config variable is not set, then I don't think the code will be able to assign a default value to the variable.

The issue here is that the devstack is not configuring the variable in the l3_agent.ini file.

Patch-set: 3