Exit after printing error about RPC config
When finding an invalid RPC configuration, devstack prints
a suitable error in the logs

    ERROR: only one rpc backend may be enabled,
           set only one of 'rabbit', 'qpid', 'zeromq'
           via ENABLED_SERVICES.

but then just carries on running, as a result the developer
is very unlikely to ever see the original error message.
Change the 'echo' to a 'die' so that it stop immediately
making the error message clearly visible.

Change-Id: If108f314e1ad2ecc79f41cbca47d08c4024a23cd