Update patch set 1
Patch Set 1:

(1 comment)

Thanks for the review Dean, as far as which lands first; the "good" thing is that I set both of these up so that if one or the other is missing we just run the same way as we always have in the past.

You're correct though, this will have zero impact on things until the Cinder change lands.

Patch-set: 1
Reviewer: Gerrit User 2243 <2243@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0