Update patch set 1
Patch Set 1:

Just chatted with Joe on infra channel. We clearly see there is a problem in devstack logs for curl'ing get-pip.py so we should get this in anyways and it would not hurt.

However Joe was mentioning that the underlying issue was that nodepool was not caching the get-pip.py in a wellknown place and hence the image was trying to do a fresh get-pip.py.

So let's please get this in and i'll see about the nodepool logs as well

Patch-set: 1
Reviewer: Gerrit User 5638 <5638@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0