Add Configuration option to configure delete namespaces
With the introduction for distributed virtual routers
the existence of namespaces in the nodes have increased
tremendously and so the namespace should be cleaned up
whenever required.

So let us provide a configurable option to configure in
the local.conf to either enable or disable the
router_delete_namespaces option for neutron.

This patch introduces "Q_DEL_NAMESPACE" which will be
used to populate the "router_delete_namespaces".

Change-Id: Ie74156a802ca9c383364b65f8b68b4c8feaa1fea
Closes-Bug: #1413777