Update patch set 1
Patch Set 1: Code-Review+1

Thanks Mike, I forgot about this.  The only question I had was making the iscsi_helper a value that could be read and set from local.conf, but we don't have anything to do that currently either so it should probably be considered as independent of this patch.

Patch-set: 1
Label: Code-Review=+1