Update patch set 1
Patch Set 1: Code-Review-1

good start, however the file needs to get downloaded in a way that we can cache it in the nodepool servers.

Look at tools/install_pip.sh

It should download the file into files/, but not if it's already there, then it should install it from there. That lets us preseed that for gate jobs (and do caching for local dev).

Patch-set: 1
Label: Code-Review=-1