Update patch set 1
Patch Set 1:

I get your point. But my little experience with openstack make me trust the code more than the documentation :) And it would be weird to have a code comment says a different thing that the code itself.

The document you pointed says "Only Ubuntu 14.04 (Trusty), Fedora 20 and CentOS/RHEL 6.5 are documented here." This means that this document (index.rst) only covers Ubuntu 14.04, Fedora 20 and CentOS/RHEL 6.5. But you want to change a code comment that enumerates the list of supported distros in devstack. Which is different from the list of distros the documentation applies to.

Patch-set: 1