Update patch set 1
Patch Set 1:

The problem is it renders here http://docs.openstack.org/developer/devstack/stack.sh.html and that doc as it is currently contradicts the rest of the documentation there. The comment - even before - effectively tells the user that it may work with other distros but these are what we're expecting to work.

There are other errors addressed here - the F18 mention needed updating regardless - but this Ubuntu part we disagree on. I'm curious to see others weigh in on whether we should call out 12.04 (in which case, shouldn't the index call that out as a valid minimum rather than 14.04?) or if this change makes sense as it is.

Patch-set: 1