Update patch set 4
Patch Set 4:

I've changed it back to 6.5. I'm not sure you can even run that without major handholding for Python 2.7 but at least it's in the realm of possibility. 6.0 seems very unlikely to work with Kilo, period.

Yes, I realize it says rhel6 in there but that level of granularity is not sufficient to distinguish the changes between 6.0 and 6.5.

As for the build breaks - it's obviously not related to this doc update so I'm working on that separately (it's not clear from the logs what's actually breaking.)

Patch-set: 4
Label: Verified=0