Update patch set 2
Patch Set 2: Code-Review-1

I agree that back-porting isn't a good idea unless there is a very compelling reason.  at least juno *was* being tested on f20, even if bit-rot is starting to set in as the world moves on.

If there's some argument that it's part of a CI and it will be monitored and tested, well that's another thing, but i'm not aware of that.

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1