Update patch set 2
Patch Set 2: Code-Review-1

There is no point to removing this.  We backport far too many things to stable branches that are not bug fixes or stabilization.  This is an indication of what platforms were supported for this release and has a historical purpose if not a functional one.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1