Update patch set 1
Patch Set 1:

Fair enough, these can all be set with the local.conf file. Maybe I misunderstood the purpose of the cinder_backends scripts, the others seem to do this same kind of thing by being a helper to simplify the configuration.

My main goal for adding this is to reduce confusion in configuring DevStack to use our Cinder backend. Right now you need to set something like: CINDER_ENABLED_BACKENDS=pure:puredriver-1 in local.conf, where the "pure" prefix means nothing (but some string is required), and the "puredriver-1" string is pretty much arbitrary, but *must* match up with default_volume_type and volume_backend_name settings for cinder to make the tests pass due to the default behavior of the cinder setup script.

With this change that CINDER_ENABLED_BACKENDS=pure:puredriver-1 takes advantage of the smarts built in to the cinder setup script by choosing to use the "pure" cinder_backend script to setup the correct backend names and volume types so you don't need to make sure all the strings match up.

So, with all that in mind, is simplification of config files something preferred or to remove complexity from the setup scripts and keep more in the config files? I can understand the benefits of either direction...

Patch-set: 1