Update patch set 1
Patch Set 1:

(2 comments)

Ian, there is a bigger plan involving venvs for the CLI clients used and the services with their own copies of the client libs.  My apologies for not communicating it outside the Nova/qa channels where we finalized this last week.  And admittedly some of it was in person at the Nova meetup that I happened to be in the Bay Area at the same time.  The venv topic started to be seriously consider in one of the sessions in Paris.  I took two stabs at it (one in the middle of the night for an icehouse breakage) and now this series.

[much of this is to record some IRC discussions that I think cleared up the overall intent]

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0