Stop installing tempest
We run tempest inside of tox so no nee to install it. By not installing
it we decouple one more thing from the requirements sync.

Without this patch, due to branchless tempest, tempest must work with
master and all stable branch global requirements.

WIP: Make installing tempest a devstack flag that is on by default, but
that we can turn off on stable/branches in devstack-gate etc. As
installing tempest makes it easier to run for development use cases.

Change-Id: I78d51f04ed01da4ce8aa0e127be028f969d3b4f8