Update patch set 1
Patch Set 1:

Hey Jordan, you are correct that this can be avoided by setting the range to a different value... but imo that feels more like a hack for using the image_list.sh tool than adding a flag to disable the check.

I'm not 100% convinced my proposed fix is the best way, but I do think it is a bug that you can't run this helper script to find a list of images if the host has a 10.0.0.x ip address. They shouldn't be related.

Patch-set: 1
Reviewer: Gerrit User 12924 <12924@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0