Use 'https' protocol by default as easier for proxies
The 'https' protocol is easily handled by proxy servers.  The 'git'
proxy, not so much.

Change the default protocol to be 'https' for git repositories.

Using 'git://' as the protocol requires more extensive configuration of
proxy servers and in particular requires usage of a SOCKS proxy.  While
'https://' is easily handled by setting the environment variable
'https_proxy'.

In limited testing the time to 'git clone' the Nova project was
comparable for both 'https://' and 'git://'

This is reverting part of the change introduced by
commit 047cac56e14552eb6b2d374a35f3a092c5f2a5d4 which changed the
URL from https://github.com/ to git://git.openstack.org/ on 7-Nov-2013.

Prior to that change 'https://' was used as the default protocol going
back to 25-Oct-2011 (commit c2b3cf7ff87cd93113a36eb9756705b905a7f6dc)

'https://' as a protocol for Git is widely used by sites like
https://github.com/ and https://android-review.googlesource.com/
https://android-review.googlesource.com/ does not support 'git://'

Change-Id: Ib59c5b5cc986dacea2c31a787c89040e7534010e
Signed-off-by: John L. Villalovos <john.l.villalovos@intel.com>